Type Checking Command
Was using this setup already, so I thought I'd add an npm script to do it more simply. It's recommended to keep track of the TypeScript version you're using with your builds too, to ensure the same results can be replicated when installing older versions of your app again down the line. Helps with debugging in the future!